requirements.txt
tornado==4.3
dj-database-url==0.3.0
Django==1.8.1
django-postgrespool==0.3.0
gunicorn==19.3.0
SQLAlchemy==1.0.4
whitenoise==1.0.6
django-toolbelt==0.0.1
static3==0.5.1
requests==2.3.0
wsgiref==0.1.2
python-telegram-bot==3.0
psycopg2==2.6